fix: revert split queue broker
