--- bifrost/bifrost_queue_broker/bifrost_queue_broker_tbr_pull/requirements.txt
@@ -1,6 +1,0 @@
-pymongo==3.11.0
-pymongocrypt==1.0.1
-urllib3==1.26.2
-python-dateutil==2.8.1
-nulltype==2.3.1
-commentjson==0.9.0
