+++ bifrost/bifrost_queue_broker/brokers/pulling/lims_pulling_broker.py
@@ -1,8 +1,11 @@
+import os, sys
 import logging
-import pymongo
 import time
-from brokers.shared import  yield_chunks
-from brokers.lims_conn import *
+import pymongo
+import threading
+from pymongo import CursorType
+from ..shared import BrokerError, yield_chunks
+from ..lims_conn import *
 from common.database import (
     coerce_dates,
     encrypt_dict,
@@ -11,32 +14,53 @@ from common.database import (
 from common.config.column_config import pii_columns
 
 # LIMS API imports
+import time
 import api_clients.lims_client
-from api_clients.lims_client.api import isolate_api
+from api_clients.lims_client.api import connections_api, isolate_api
 from api_clients.lims_client.models import (
     IsolateGetRequest,
+    IsolateGetResponse,
+    ConnectionCreateRequest,
+    ConnectionCreateResponse,
 )
 
-class LIMSPullingBrokerSync():
+
+class LIMSPullingBroker(threading.Thread):
     def __init__(
-        self, lims_col_name, analysis_view_col_name, db
+        self, data_lock, lims_col_name, analysis_view_col_name, thread_timeout, db
     ):
-        super(LIMSPullingBrokerSync, self).__init__()
+        super(LIMSPullingBroker, self).__init__()
+        self.data_lock = data_lock
         self.db = db
         self.analysis_col = db[analysis_view_col_name]
         self.metadata_col = db[lims_col_name]
+        self.stop = threading.Event()
+        self.thread_timeout = thread_timeout
         self.broker_name = "LIMS Pulling broker"
         _, enc = get_connection(with_enc=True)
         self.encryption_client = enc
 
+    def stop(self):
+        self.stop.set()
+
     def run(self):
         logging.info(f"Started {self.broker_name} thread.")
         interval = 60 * 10  # 10 minutes
-        start_time = time.time()
         first_run = True
-        while first_run or (time.time() - start_time) < interval:
+        while first_run or not self.stop.wait(interval):
             first_run = False
-            self.run_sync_job()
+            thread_acquired = self.data_lock.acquire(timeout=self.thread_timeout)
+            if not thread_acquired:
+                logging.warning(
+                    f"{self.broker_name} Failed to acquire thread before {self.thread_timeout} seconds timeout"
+                )
+                continue
+            try:
+                self.run_sync_job()
+            except:
+                raise
+            finally:
+                self.data_lock.release()
 
     def run_sync_job(self):
         batch_size = 200
@@ -89,28 +113,37 @@ class LIMSPullingBrokerSync():
     def update_isolate_metadata(self, api_client, element_batch):
         api_instance = isolate_api.IsolateApi(api_client)
         transformed_batch = []
-        for element in element_batch:
-            isolate_get_req = IsolateGetRequest(isolate_id=element["isolate_id"])
+        try:
+            for element in element_batch:
+                isolate_get_req = IsolateGetRequest(isolate_id=element["isolate_id"])
 
-            api_response = api_instance.post_actions_get_isolate(
-                isolate_get_request=isolate_get_req
-            )
-            if (
-                "output" in api_response
-                and "sapresponse" in api_response.output
-                and api_response.output.sapresponse.success
-            ):
-                transformed_batch.append(transform_lims_metadata(api_response))
+                api_response = api_instance.post_actions_get_isolate(
+                    isolate_get_request=isolate_get_req
+                )
+                try:
+                    if (
+                        "output" in api_response
+                        and "sapresponse" in api_response.output
+                        and api_response.output.sapresponse.success
+                    ):
+                        transformed_batch.append(transform_lims_metadata(api_response))
+                except Exception as e:
+                    logging.debug(f"Skipping isolate: {element['isolate_id']}")
 
-        bulk_update_queries = self.upsert_lims_metadata_batch(transformed_batch)
-        update_count = 0
-        if len(bulk_update_queries) > 0:
-            bulk_result = self.metadata_col.bulk_write(
-                bulk_update_queries, ordered=False
-            )
-            update_count = bulk_result.upserted_count + bulk_result.modified_count
+            bulk_update_queries = self.upsert_lims_metadata_batch(transformed_batch)
+            update_count = 0
+            if len(bulk_update_queries) > 0:
+                bulk_result = self.metadata_col.bulk_write(
+                    bulk_update_queries, ordered=False
+                )
+                update_count = bulk_result.upserted_count + bulk_result.modified_count
 
-        return update_count
+            return update_count
+
+        except Exception as e:
+            print("Exception when updating metadata: %s\n" % e)
+
+        return len(element_batch)
 
     def upsert_lims_metadata_batch(self, metadata_batch):
         result = []
