+++ bifrost/bifrost_queue_broker/test-requirements.txt
@@ -1,0 +1,3 @@
+pytest~=4.6.7 # needed for python 2.7+3.4
+pytest-cov>=2.8.1
+pytest-randomly==1.2.3 # needed for python 2.7+3.4
